fix the tile size + draw a minimap +...

+++ Projects/cub3d/functions/configuration/ft_set_array2d.c
@@ -5,10 +5,11 @@ void  ft_set_array_2d()
     config.columns = config.max_length;
     config.rows = ft_lst_size(config.head);
 
-    config.tile_width =(config.width / config.columns);
-    //config.TILE_SIZE_width = 90;
-    config.tile_height =(float)(config.height /config.rows);
-    //config.TILE_SIZE_height = 90;
+    config.tile_width =(config.width/config.columns);
+   
+    //config.tile_width = 40;
+    config.tile_height =(config.height /config.rows);
+    //config.tile_height = 40;
     
     int r = 0;
